ECS and SBT stack dependencies delete
