+++ server/application/Dockerfile.product
@@ -9,5 +9,16 @@ RUN yarn build product
 FROM public.ecr.aws/bitnami/node:20
 WORKDIR /app
 COPY --from=build /app ./
+
+########################WORKSHOP-TEST##############################
+RUN apt-get update && \
+    apt-get install -y \
+    unzip \
+    && curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip" \
+    && unzip awscliv2.zip \
+    && ./aws/install \
+    && rm -rf awscliv2.zip 
+########################WORKSHOP-TEST##############################
+
 EXPOSE 3010
 CMD ["npm", "run", "start:product"]
