+++ server/application/reverseproxy/nginx.conf
@@ -32,7 +32,7 @@ http {
 
       # Only requests matching the expectations will
       # get sent to the application server
-      proxy_pass http://orders-api.prod.sc:3010;
+      proxy_pass http://orders-api.${NAMESPACE}.sc:3010;
       proxy_http_version 1.1;
       proxy_set_header Upgrade $http_upgrade;
       proxy_set_header Connection 'upgrade';
@@ -48,7 +48,7 @@ http {
 
       # Only requests matching the expectations will
       # get sent to the application server
-      proxy_pass http://products-api.prod.sc:3010;
+      proxy_pass http://products-api.${NAMESPACE}.sc:3010;
       proxy_http_version 1.1;
       proxy_set_header Upgrade $http_upgrade;
       proxy_set_header Connection 'upgrade';
@@ -64,7 +64,7 @@ http {
 
       # Only requests matching the expectations will
       # get sent to the application server
-      proxy_pass http://users-api.prod.sc:3010;
+      proxy_pass http://users-api.${NAMESPACE}.sc:3010;
       proxy_http_version 1.1;
       proxy_set_header Upgrade $http_upgrade;
       proxy_set_header Connection 'upgrade';
