+++ server/lib/bootstrap-template/control-plane-stack.ts
@@ -1,15 +1,18 @@
 import * as cdk from 'aws-cdk-lib';
 import { type Construct } from 'constructs';
-import { StaticSiteDistro } from './static-site-distro';
 import path = require('path');
 import { StaticSite } from './static-site';
 import { ControlPlaneNag } from '../cdknag/control-plane-nag';
 import { addTemplateTag } from '../utilities/helper-functions';
 import * as sbt from '@cdklabs/sbt-aws';
+import { StaticSiteDistro } from '../shared-infra/static-site-distro';
 
 interface ControlPlaneStackProps extends cdk.StackProps {
   systemAdminRoleName: string
   systemAdminEmail: string
+  accessLogsBucket: cdk.aws_s3.Bucket
+  distro: StaticSiteDistro
+  adminSiteUrl: string
 }
 
 export class ControlPlaneStack extends cdk.Stack {
@@ -22,25 +25,9 @@ export class ControlPlaneStack extends cdk.Stack {
   constructor (scope: Construct, id: string, props: ControlPlaneStackProps) {
     super(scope, id, props);
     addTemplateTag(this, 'ControlPlaneStack');
-    
-    const accessLogsBucket = new cdk.aws_s3.Bucket(this, 'AccessLogsBucket', {
-      enforceSSL: true,
-      autoDeleteObjects: true,
-      accessControl: cdk.aws_s3.BucketAccessControl.LOG_DELIVERY_WRITE,
-      removalPolicy: cdk.RemovalPolicy.DESTROY
-    });
-
-    const distro = new StaticSiteDistro(this, 'StaticSiteDistro', {
-      allowedMethods: ['GET', 'HEAD', 'OPTIONS'],
-      accessLogsBucket
-    });
-
-    this.adminSiteUrl = `https://${distro.cloudfrontDistribution.domainName}`;
 
     const cognitoAuth = new sbt.CognitoAuth(this, 'CognitoAuth', {
-      // Avoid checking scopes for API endpoints. Done only for testing purposes.
-      // setAPIGWScopes: false,
-      controlPlaneCallbackURL: this.adminSiteUrl
+      controlPlaneCallbackURL: props.adminSiteUrl
     });
 
     const controlPlane = new sbt.ControlPlane(this, 'controlplane-sbt', {
@@ -67,13 +54,17 @@ export class ControlPlaneStack extends cdk.Stack {
       issuer: this.auth.tokenEndpoint,
       apiUrl: this.regApiGatewayUrl,
       wellKnownEndpointUrl: this.auth.wellKnownEndpointUrl,
-      distribution: distro.cloudfrontDistribution,
-      appBucket: distro.siteBucket,
-      accessLogsBucket
+      distribution: props.distro.cloudfrontDistribution,
+      appBucket: props.distro.siteBucket,
+      accessLogsBucket: props.accessLogsBucket,
+      env: {
+        account: this.account,
+        region: this.region
+      }
     });
     
     new cdk.CfnOutput(this, 'adminSiteUrl', {
-      value: this.adminSiteUrl
+      value: props.adminSiteUrl
     });
 
     new ControlPlaneNag(this, 'controlplane-nag');
