+++ server/lib/bootstrap-template/core-appplane-stack.ts
@@ -1,23 +1,26 @@
 import * as cdk from 'aws-cdk-lib';
+import * as fs from 'fs';
 import { type Construct } from 'constructs';
 import { Table, AttributeType } from 'aws-cdk-lib/aws-dynamodb';
 import { PolicyDocument } from 'aws-cdk-lib/aws-iam';
-import * as fs from 'fs';
-import { CoreAppPlaneNag } from '../cdknag/core-app-plane-nag';
 import { addTemplateTag } from '../utilities/helper-functions';
-import { StaticSiteDistro } from './static-site-distro';
+import { StaticSiteDistro } from '../shared-infra/static-site-distro';
 import path = require('path');
 import { StaticSite } from './static-site';
+import { CoreAppPlaneNag } from '../cdknag/core-app-plane-nag';
 import * as sbt from '@cdklabs/sbt-aws';
 
 interface CoreAppPlaneStackProps extends cdk.StackProps {
   eventManager: sbt.IEventManager
   systemAdminEmail: string
   regApiGatewayUrl: string
+  distro: StaticSiteDistro
+  appSiteUrl: string
+  accessLogsBucket: cdk.aws_s3.Bucket
+  tenantMappingTable: Table
 }
 
 export class CoreAppPlaneStack extends cdk.Stack {
-  public readonly tenantMappingTable: Table;
   public readonly appBucket: cdk.aws_s3.Bucket;
   public readonly appSiteUrl: string;
 
@@ -25,25 +28,7 @@ export class CoreAppPlaneStack extends cdk.Stack {
     super(scope, id, props);
     addTemplateTag(this, 'CoreAppPlaneStack');
 
-    const accessLogsBucket = new cdk.aws_s3.Bucket(this, 'AccessLogsBucket', {
-      enforceSSL: true,
-      autoDeleteObjects: true,
-      accessControl: cdk.aws_s3.BucketAccessControl.LOG_DELIVERY_WRITE,
-      removalPolicy: cdk.RemovalPolicy.DESTROY
-    });
-
-    const distro = new StaticSiteDistro(this, 'StaticSiteDistro', {
-      allowedMethods: ['GET', 'HEAD', 'OPTIONS'],
-      accessLogsBucket
-    });
-
-    this.appSiteUrl = `https://${distro.cloudfrontDistribution.domainName}`;
-
     const systemAdminEmail = props.systemAdminEmail;
-
-    this.tenantMappingTable = new Table(this, 'TenantMappingTable', {
-      partitionKey: { name: 'tenantId', type: AttributeType.STRING }
-    });
 
     const provisioningScriptJobProps = {
       permissions: PolicyDocument.fromJson(
@@ -105,7 +90,7 @@ export class CoreAppPlaneStack extends cdk.Stack {
       outgoingEvent: sbt.DetailType.DEPROVISION_SUCCESS,
       incomingEvent: sbt.DetailType.OFFBOARDING_REQUEST,
       scriptEnvironmentVariables: {
-        TENANT_STACK_MAPPING_TABLE: this.tenantMappingTable.tableName,
+        TENANT_STACK_MAPPING_TABLE: props.tenantMappingTable.tableName,
         // CDK_PARAM_SYSTEM_ADMIN_EMAIL is required because as part of deploying the bootstrap-template
         // the control plane is also deployed. To ensure the operation does not error out, this value
         // is provided as an env parameter.
@@ -132,13 +117,17 @@ export class CoreAppPlaneStack extends cdk.Stack {
       assetDirectory: path.join(__dirname, '../../../client/Application'),
       production: true,
       apiUrl: props.regApiGatewayUrl,
-      distribution: distro.cloudfrontDistribution,
-      appBucket: distro.siteBucket,
-      accessLogsBucket
+      distribution: props.distro.cloudfrontDistribution,
+      appBucket: props.distro.siteBucket,
+      accessLogsBucket: props.accessLogsBucket,
+      env: {
+        account: this.account,
+        region: this.region
+      }
     });
 
     new cdk.CfnOutput(this, 'appSiteUrl', {
-      value: this.appSiteUrl
+      value: props.appSiteUrl
     });
 
     new CoreAppPlaneNag(this, 'CoreAppPlaneNag');
